streamVideo.js -> separate js files
